pyproject.toml
[project]
name = "autoloader-framework-pydab"
version = "0.1.0"
description = "Databricks Asset Bundle for Autoloader Jobs with DLT/Lakeflow Declarative Pipelines"
requires-python = ">=3.10"
dependencies = [
    "databricks-bundles>=0.248.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "resources"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
] 